Move AxisOrder to appropriate package
