+++ src/main/java/org/janelia/saalfeldlab/paintera/data/HasAxisOrder.java
@@ -1,6 +1,6 @@
 package org.janelia.saalfeldlab.paintera.data;
 
-import org.janelia.saalfeldlab.paintera.data.mask.AxisOrder;
+import org.janelia.saalfeldlab.paintera.data.axisorder.AxisOrder;
 
 public interface HasAxisOrder {
