+++ src/main/java/org/janelia/saalfeldlab/paintera/data/axisorder/AxisOrder.java
@@ -1,12 +1,9 @@
-package org.janelia.saalfeldlab.paintera.data.mask;
+package org.janelia.saalfeldlab.paintera.data.axisorder;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
-import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
 public enum AxisOrder {
